34. A Small JavaScript Dropdown Detour
We next need to make that "Categories" dropdown on the home page function as expected. I hate to break it to you, but we'll need to reach for a bit of JavaScript to make this work. Don't worry: I'll make this as painless as possible by pulling in the excellent Alpine.js library. Let's get through this, and we'll jump back into some Laravel-specific topics!