40. Extract a Category Dropdown Blade Component
Have you noticed that each route needs to pass a collection of categories to the posts view? If you take a look, that variable is only ever referenced as part of the main category dropdown. So with that in mind, what if we created a dedicated x-category-dropdown component that could be responsible for fetching any data that the dropdown requires? Let's figure out how to allow for that in this episode.