44. Laughably Simple Pagination
We're currently fetching all posts from the database and rendering them as a grid on the home page. But what happens down the line when you have, say, five hundred blog posts? That's a bit too costly to render. The solution is to leverage pagination - and luckily, Laravel does all the work for you. Come see for yourself.